+++ autorest/retriablerequest_1.7.go
@@ -48,7 +48,7 @@ func (rr *RetriableRequest) Prepare() (err error) {
 	return err
 }
 
-func removeRequestBody(req *http.Request) {
+func RemoveRequestBody(req *http.Request) {
 	req.Body = nil
 	req.ContentLength = 0
 }
