+++ autorest/retriablerequest_1.8.go
@@ -59,7 +59,7 @@ func (rr *RetriableRequest) Prepare() (err error) {
 	return err
 }
 
-func removeRequestBody(req *http.Request) {
+func RemoveRequestBody(req *http.Request) {
 	req.Body = nil
 	req.GetBody = nil
 	req.ContentLength = 0
